feat: add pathString prop to LogicGroup, FilterTypeSwitch, and Rule components
The pathString prop is added to the LogicGroup, FilterTypeSwitch, and Rule components. This prop is used to keep track of the current path of the component in the filter hierarchy. It is updated dynamically based on the path and the filter structure. This change allows for more accurate tracking and manipulation of the filter structure.